requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.30.*
SQLAlchemy[asyncio]==2.0.*
asyncpg==0.29.*
python-dotenv==1.0.*
PyJWT==2.9.*
bcrypt==4.2.*
email-validator==2.1.*
